Fix benchmark module to make it work with latest version of the universe framework
