+++ it/src/test/java/org/corfudb/universe/GenericIntegrationTest.java
@@ -8,9 +8,6 @@ import org.junit.Before;
 import org.junit.Rule;
 import org.junit.rules.TestName;
 
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
 import java.util.function.Consumer;
 
 /**
@@ -21,7 +18,7 @@ import java.util.function.Consumer;
  * to manage your initialization and deployment process in the universe.
  */
 public abstract class GenericIntegrationTest {
-    private static final AppUtil APP_UTIL = new AppUtil();
+    private static final UniverseAppUtil APP_UTIL = new UniverseAppUtil();
 
     @Rule
     public TestName test = new TestName();
@@ -43,25 +40,4 @@ public abstract class GenericIntegrationTest {
         return universeManager.workflow(action);
     }
 
-    private static class AppUtil {
-
-        /**
-         * Provides a current version of this project. It parses the version from pom.xml
-         *
-         * @return maven/project version
-         */
-        public String getAppVersion() {
-            String version = System.getProperty("project.version");
-            if (version != null && !version.isEmpty()) {
-                return version;
-            }
-
-            try {
-                Path path = Paths.get(ClassLoader.getSystemResource("corfu.version").toURI());
-                return new String(Files.readAllBytes(path));
-            } catch (Exception e) {
-                throw new IllegalStateException("Corfu version file not found");
-            }
-        }
-    }
 }
